+++ je-kalender.php
@@ -3,14 +3,14 @@
 /**
  * Plugin Name: JE Kalender
  * Description: Google Kalender Integration mit Leaflet-Kartenanzeige für Veranstaltungen.
- * Version: 1.1.2
+ * Version: 1.1.3
  * Author: Jan Erbert
  */
 
 defined('ABSPATH') || exit;
 
 define('JE_KALENDER_PLUGIN_FILE', __FILE__);
-define('JE_KALENDER_PLUGIN_VERSION', '1.1.2');
+define('JE_KALENDER_PLUGIN_VERSION', '1.1.3');
 
 // Pluginmodule laden
 require_once plugin_dir_path(__FILE__) . 'includes/functions.php';